# Fixed and upgraded benchmarking