--- a/benches/compression.rs
+++ b/benches/compression.rs
@@ -1,21 +1,23 @@
 use cratesfyi::storage::{compress, decompress, CompressionAlgorithm};
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use criterion::{black_box, criterion_group, criterion_main, Criterion, Throughput};
 
 const ALGORITHM: CompressionAlgorithm = CompressionAlgorithm::Zstd;
 
-pub fn criterion_benchmark(c: &mut Criterion) {
+pub fn regex_capture_matches(c: &mut Criterion) {
     // this isn't a great benchmark because it only tests on one file
     // ideally we would build a whole crate and compress each file, taking the average
     let html = std::fs::read_to_string("benches/struct.CaptureMatches.html").unwrap();
     let html_slice = html.as_bytes();
-    c.bench_function("compress regex html", |b| {
-        b.iter(|| compress(black_box(html_slice, ALGORITHM)))
-    });
-    let compressed = compress(html_slice, ALGORITHM).unwrap();
-    c.bench_function("decompress regex html", |b| {
-        b.iter(|| decompress(black_box(compressed.as_slice()), ALGORITHM))
-    });
+
+    c.benchmark_group("regex html")
+        .throughput(Throughput::Bytes(html_slice.len() as u64))
+        .bench_function("compress", |b| {
+            b.iter(|| compress(black_box(html_slice), ALGORITHM));
+        })
+        .bench_function("decompress", |b| {
+            b.iter(|| decompress(black_box(html_slice), ALGORITHM, 5 * 1024 * 1024));
+        });
 }
 
-criterion_group!(compression, criterion_benchmark);
+criterion_group!(compression, regex_capture_matches);
 criterion_main!(compression);