The compression benchmark was broken, I fixed it and upgraded it with throughput

## Before

```
compress regex html     time:   [1.9454 ms 1.9615 ms 1.9789 ms]
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe

decompress regex html   time:   [104.77 us 105.41 us 106.12 us]
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

parse regex html        time:   [6.2537 ms 6.2975 ms 6.3477 ms]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```

## After

```
regex html/compress     time:   [1.9521 ms 1.9586 ms 1.9653 ms]
                        thrpt:  [52.237 MiB/s 52.417 MiB/s 52.591 MiB/s]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
regex html/decompress   time:   [4.6433 us 4.6499 us 4.6575 us]
                        thrpt:  [21.526 GiB/s 21.561 GiB/s 21.591 GiB/s]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe

regex html/extract head and body
                        time:   [6.4282 ms 6.4453 ms 6.4637 ms]
                        thrpt:  [15.883 MiB/s 15.928 MiB/s 15.971 MiB/s]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```